Fix test warnings

+++ trajopt_utils/include/trajopt_utils/basic_array.hpp
@@ -42,10 +42,10 @@ struct BasicArray
   }
   std::vector<T> rblock(int startRow, int startCol, int nCol) const
   {
-    std::vector<T> out(nCol);
+    std::vector<T> out(static_cast<size_t>(nCol));
     for (int iCol = 0; iCol < nCol; ++iCol)
     {
-      out[iCol] = at(startRow, iCol + startCol);
+      out[static_cast<size_t>(iCol)] = at(static_cast<size_t>(startRow), static_cast<size_t>(iCol + startCol));
     }
     return out;
   }
